requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
huggingface-hub==0.25.2
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pydantic==2.9.2
pydantic-core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
supervision==0.24.0
sympy==1.13.3
timm==1.0.9
tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
typing-extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
